Cache: solve the split-package problem
- move the classes from the quarkus-cache-runtime-spi artifact to
quarkus-cache
- get rid of the quarkus-cache-runtime-spi artifact
- move CacheManagerInfoBuildItem to the quarkus-cache-deployment
