+++ extensions/cache/pom.xml
@@ -19,7 +19,6 @@
         <module>deployment</module>
         <module>deployment-spi</module>
         <module>runtime</module>
-        <module>runtime-spi</module>
         <module>runtime-dev</module>
     </modules>
 </project>
