+++ extensions/cache/runtime/src/main/java/io/quarkus/cache/CacheManagerInfo.java
