+++ extensions/redis-cache/deployment/src/main/java/io/quarkus/cache/redis/deployment/RedisCacheProcessor.java
@@ -25,8 +25,8 @@ import org.jboss.logging.Logger;
 import io.quarkus.arc.deployment.UnremovableBeanBuildItem;
 import io.quarkus.cache.CompositeCacheKey;
 import io.quarkus.cache.deployment.CacheDeploymentConstants;
+import io.quarkus.cache.deployment.CacheManagerInfoBuildItem;
 import io.quarkus.cache.deployment.CacheNamesBuildItem;
-import io.quarkus.cache.deployment.spi.CacheManagerInfoBuildItem;
 import io.quarkus.cache.redis.runtime.RedisCacheBuildRecorder;
 import io.quarkus.cache.redis.runtime.RedisCacheBuildTimeConfig;
 import io.quarkus.cache.redis.runtime.RedisCachesBuildTimeConfig;
